pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "readbrain"
version = "0.1.0"
description = "AI-powered personal reading knowledge base"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "openai>=1.40.0",
    "pydantic>=2.8.0",
    "pyyaml>=6.0.2",
    "python-frontmatter>=1.1.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
]

[project.scripts]
readbrain = "app.cli:main"
